move instance specific hooks to instance.ts hooks collection (probably needs to be split)
