+++ app/canvas/shared/use-track-hovered-element.ts
@@ -40,16 +40,3 @@ export const useTrackHoveredElement = () => {
     };
   }, [rootInstance, setHoveredElement]);
 };
-
-export const useTrackHoveredInstance = () => {
-  const [rootInstance] = useRootInstance();
-  const [hoveredElement] = useHoveredElement();
-  const [, setHoveredInstance] = useHoveredInstance();
-
-  useEffect(() => {
-    if (rootInstance === undefined || hoveredElement === undefined) return;
-    const instance = findInstanceById(rootInstance, hoveredElement.id);
-    if (instance === undefined) return;
-    setHoveredInstance(instance);
-  }, [rootInstance, hoveredElement, setHoveredInstance]);
-};
